Use `scrollToSelector` as prop
