+++ client/landing/stepper/declarative-flow/internals/steps-repository/pattern-assembler/pattern-assembler-preview.tsx
@@ -4,7 +4,7 @@ import { Spinner } from '@wordpress/components';
 import { useSelect } from '@wordpress/data';
 import classnames from 'classnames';
 import { useTranslate } from 'i18n-calypso';
-import { useEffect, useState, useRef } from 'react';
+import { useEffect, useState } from 'react';
 import ReactDOM from 'react-dom';
 import WebPreview from 'calypso/components/web-preview/content';
 import { recordTracksEvent } from 'calypso/lib/analytics/tracks';
@@ -20,17 +20,16 @@ interface Props {
 	header: Pattern | null;
 	sections: Pattern[];
 	footer: Pattern | null;
+	scrollToSelector: string | null;
 }
 
-const PatternAssemblerPreview = ( { header, sections = [], footer }: Props ) => {
+const PatternAssemblerPreview = ( { header, sections = [], footer, scrollToSelector }: Props ) => {
 	const locale = useLocale();
 	const translate = useTranslate();
 	const site = useSite();
 	const [ webPreviewFrameContainer, setWebPreviewFrameContainer ] = useState< Element | null >(
 		null
 	);
-	const [ scrollToSelector, setScrollToSelector ] = useState< string | null >( null );
-	const prevSectionsRef: { current: undefined | Pattern[] } = useRef();
 	const hasSelectedPatterns = header || sections.length > 0 || footer;
 	const selectedDesign = useSelect( ( select ) => select( ONBOARD_STORE ).getSelectedDesign() );
 
@@ -50,44 +49,6 @@ const PatternAssemblerPreview = ( { header, sections = [], footer }: Props ) => 
 	useEffect( () => {
 		setWebPreviewFrameContainer( document.querySelector( '.web-preview__frame-wrapper' ) );
 	}, [] );
-
-	useEffect( () => {
-		setScrollToSelector( null );
-	}, [ header ] );
-
-	useEffect( () => {
-		setScrollToSelector( footer ? 'footer' : null );
-	}, [ footer ] );
-
-	useEffect( () => {
-		const prevSections = prevSectionsRef.current;
-		const setScrollToSelectorByPosition = ( position: number ) => {
-			const patternPosition = header ? position + 1 : position;
-			setScrollToSelector( `.entry-content > .wp-block-group:nth-child( ${ patternPosition } )` );
-		};
-
-		if ( ! prevSections || ! sections.length ) {
-			prevSectionsRef.current = sections;
-			setScrollToSelector( null );
-			return;
-		}
-
-		if ( prevSections.length === sections.length ) {
-			// Ordered
-			const orderedSectionIndex = sections.findIndex( ( section, index ) => {
-				return section.id !== prevSections[ index ].id;
-			} );
-			setScrollToSelectorByPosition( orderedSectionIndex + 1 );
-		} else if ( sections.length > prevSections.length ) {
-			// Added
-			setScrollToSelectorByPosition( sections.length );
-		} else {
-			// Removed
-			setScrollToSelector( null );
-		}
-
-		prevSectionsRef.current = sections;
-	}, [ sections, header ] );
 
 	return (
 		<div
